feat(org): add editing protection to only OWNERS, invited people won't be able to edit

+++ src/code/services/orgService.ts
@@ -10,6 +10,11 @@ const client = new MailtrapClient({
 	token: MAIL_TOKEN,
 });
 
+const accessLevels = {
+	OWNER: "OWNER",
+	VISITOR: "VISITOR",
+};
+
 interface memberObj {
 	name: string;
 	email: string;
@@ -19,7 +24,7 @@ interface memberObj {
 interface orgObj {
 	name: string;
 	description: string;
-	organization_members: [memberObj];
+	organization_members: memberObj[];
 }
 
 async function getMongoCollection(collectionName: string) {
@@ -28,8 +33,18 @@ async function getMongoCollection(collectionName: string) {
 	return collection;
 }
 
-async function createNewOrganization(data: orgObj) {
+async function createNewOrganization(
+	data: orgObj,
+	name: string,
+	email: string,
+) {
 	const collection = await getMongoCollection("organizations");
+	data.organization_members = [];
+	data.organization_members.push({
+		name,
+		email,
+		access_level: accessLevels.OWNER,
+	});
 	const record = await collection.insertOne(data);
 	return record.insertedId;
 }
@@ -49,9 +64,15 @@ async function getSingleOrg(orgId: string) {
 	return record;
 }
 
-async function updateOrg(orgId: string, data: orgObj) {
+async function updateOrg(orgId: string, data: orgObj, userEmail: string) {
 	const collection = await getMongoCollection("organizations");
 	const objectId = new ObjectId(orgId);
+	const organization = await collection.findOne({ _id: objectId });
+	const owner = organization.organization_members.find(
+		(user) => user.access_level == accessLevels.OWNER,
+	);
+	if (owner.email != userEmail)
+		throw AppError.unauthorized("You cannot edit this organization");
 	const record = await collection.findOneAndUpdate(
 		{ _id: objectId },
 		{ $set: data },
@@ -59,11 +80,32 @@ async function updateOrg(orgId: string, data: orgObj) {
 	return record;
 }
 
-async function deleteOrg(orgId: string) {
+async function deleteOrg(orgId: string, userEmail: string) {
 	const collection = await getMongoCollection("organizations");
 	const objectId = new ObjectId(orgId);
+	const organization = await collection.findOne({ _id: objectId });
+	const owner = organization.organization_members.find(
+		(user) => user.access_level == accessLevels.OWNER,
+	);
+	if (owner.email != userEmail)
+		throw AppError.unauthorized("You cannot delete this organization");
 	await collection.findOneAndDelete({ _id: objectId });
 	return;
+}
+
+async function addVisitorToOrganization(orgId: string, receiverEmail: string) {
+	const collection = await getMongoCollection("organizations");
+	await collection.findOneAndUpdate(
+		{ _id: new ObjectId(orgId) },
+		{
+			$push: {
+				organization_members: {
+					email: receiverEmail,
+					access_level: accessLevels.VISITOR,
+				},
+			},
+		},
+	);
 }
 
 async function sendInvite(
@@ -92,6 +134,8 @@ async function sendInvite(
 		text: `Congrats you are invited to join our organization \"${record.name}\" with for sending test email with Mailtrap!`,
 		category: "Mail Test",
 	});
+
+	addVisitorToOrganization(orgId, receiverEmail);
 }
 
 export = {
